third_party/WebKit/LayoutTests/platform/win/external/wpt/user-timing/test_user_timing_mark_and_measure_exception_when_invoke_without_parameter-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.mark() threw an exception when invoke without a parameter. 
PASS window.performance.mark() threw a TYPE_ERR exception when invoke without a parameter. 
PASS window.performance.measure() threw an exception when invoke without a parameter. 
PASS window.performance.measure() threw a TYPE_ERR exception when invoke without a parameter. 
Harness: the test ran to completion.
